+++ web/packages/pyxis-components/src/public/EthosGrid/index.ts
@@ -1,1 +1,0 @@
-export { EthosGrid } from './EthosGrid';