+++ web/packages/pyxis-components/src/public/FindUsBlock/index.ts
@@ -1,1 +1,0 @@
-export { FindUsBlock } from './FindUsBlock';