+++ web/packages/pyxis-components/src/public/organisms/AboutHero/AboutHero.tsx
@@ -1,5 +1,5 @@
 import { clsx } from 'clsx';
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import './AboutHero.css';
 
 export type AboutHeroProps = {