+++ web/packages/pyxis-components/src/public/organisms/CollectiveList/CollectiveList.tsx
@@ -1,5 +1,5 @@
 import { clsx } from 'clsx';
-import { pyxisPart } from '../../utils/parts';
+import { pyxisPart } from '../../../utils/parts';
 import './CollectiveList.css';
 
 export type CollectiveListProps = { className?: string };