+++ web/packages/pyxis-components/src/public/organisms/CollectiveList/index.ts
@@ -1,0 +1,2 @@
+export { CollectiveList } from './CollectiveList';
+export type { CollectiveListProps } from './CollectiveList';